Rename CommandFiles -> Commands, and restrict variants of acceptable commandfile names to just the one preferred option.
