+++ lib/Drush/Commands/ExampleCommands.php
@@ -1,5 +1,5 @@
 <?php
-namespace Drush\CommandFiles;
+namespace Drush\Commands;
 
 /**
  * @file
